requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
algoliasearch-django>=3.0,<4.0